+++ src/main/java/com/develhope/spring/admin/adminControllerResponse/ShowEarningsInPeriodRangeResponse.java
@@ -5,11 +5,13 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
 public class ShowEarningsInPeriodRangeResponse {
 
     private String message;
-    private Integer earnings;
+    private BigDecimal earnings;
 }
